refactor: decompose agents.rs into submodule directory
Break down the 3,098-line agents.rs monolith into a well-organized
module directory with focused, single-responsibility files:

  src/agents/
  ├── mod.rs          # Module declarations and public re-exports
  ├── types.rs        # AgentType, strategy enums, AgentPolyfillConfig
  ├── definition.rs   # AgentDefinition struct and built-in constructors
  ├── manager.rs      # AgentVersion, AgentManager, version operations
  ├── install.rs      # Platform detection, binary install, asset picker
  ├── custom_cli.rs   # AddCustomAgentArgs and custom agent CLI functions
  └── tests.rs        # All 55 unit tests (unchanged logic)

This is a pure structural refactoring with zero API changes:
- All public and pub(crate) items are re-exported through mod.rs
- External code continues to use `crate::agents::{AgentType, ...}`
- All 728 tests pass, cargo clippy and cargo fmt are clean
- Three private methods changed to pub(super) for test access only

No functional changes.
